% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_data.R
\name{get_api_company}
\alias{get_api_company}
\title{GET 'company' data from the SpaceX API}
\usage{
get_api_company()
}
\value{
A list of the 'company' data points.
}
\description{
GET 'company' data from the SpaceX API
}
