% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_data.R
\name{get_api_data}
\alias{get_api_data}
\title{GET data from the SpaceX API}
\usage{
get_api_data(url)
}
\arguments{
\item{url}{The url of the api.}
}
\value{
An R object containing the data.
}
\description{
GET data from the SpaceX API
}
