% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_data.R
\name{get_api_launches}
\alias{get_api_launches}
\title{GET 'launches' data from the SpaceX API}
\usage{
get_api_launches()
}
\value{
A list of the 'launches' data points.
}
\description{
GET 'launches' data from the SpaceX API
}
