% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_data.R
\name{get_api_launchpads}
\alias{get_api_launchpads}
\title{GET 'launchpads' data from the SpaceX API}
\usage{
get_api_launchpads()
}
\value{
A list of the 'launchpads' data points.
}
\description{
GET 'launchpads' data from the SpaceX API
}
