% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_data.R
\name{get_api_rockets}
\alias{get_api_rockets}
\title{GET 'rockets' data from the SpaceX API}
\usage{
get_api_rockets()
}
\value{
A list of the 'rockets' data points.
}
\description{
GET 'rockets' data from the SpaceX API
}
