% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(..., options = list())
}
\arguments{
\item{...}{A series of options to be used inside the app.}

\item{options}{A list of options.}
}
\description{
Run the Shiny Application
}
