% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tabPanel_about.R
\name{tags_address}
\alias{tags_address}
\title{Create HTML for an Address}
\usage{
tags_address(lst_company)
}
\arguments{
\item{lst_company}{A reactive list of company information.}
}
\value{
A list of HTML tags.
}
\description{
Create HTML for an Address
}
