% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tabPanel_about.R
\name{tags_links}
\alias{tags_links}
\title{Create HTML for Links}
\usage{
tags_links(lst_company)
}
\arguments{
\item{lst_company}{A reactive list of company information.}
}
\value{
A list of HTML tags.
}
\description{
Create HTML for Links
}
